+++ config.js
@@ -1,0 +1,10 @@
+const maxHeight = 10000
+const seaLevel = 6000
+const landHeight = maxHeight - seaLevel
+
+
+module.exports = {
+  maxHeight,
+  seaLevel,
+  landHeight
+}
